refactor: break components and functions
