+++ front-end/src/components/ProfileDialog.tsx
@@ -1,4 +1,4 @@
-import { programmingLanguages, experienceLevels } from "@/data";
+import { programmingLanguages, experienceLevels } from "@/lib/data";
 import { Save } from "lucide-react";
 import { Button } from "./ui/button";
 import {
