+++ front-end/src/components/RepositoryCard.tsx
@@ -1,5 +1,3 @@
-import { Badge } from "@/components/ui/badge";
-import { Button } from "@/components/ui/button";
 import {
   Card,
   CardHeader,
@@ -7,86 +5,9 @@ import {
   CardContent,
   CardFooter,
 } from "@/components/ui/card";
-import {
-  BookOpen,
-  Heart,
-  Star,
-  GitFork,
-  AlertCircle,
-  Clock,
-} from "lucide-react";
-import { Repository } from "../pages/public/LandingPage/SearchSection";
-
-const icons: Record<string, { icon: string; color: string }> = {
-  javascript: {
-    icon: "🌟",
-    color: "#F7DF1E",
-  },
-  typescript: {
-    icon: "🌀",
-    color: "#3178C6",
-  },
-  python: {
-    icon: "🐍",
-    color: "#3776AB",
-  },
-  java: {
-    icon: "☕",
-    color: "#007396",
-  },
-  c: {
-    icon: "🔵",
-    color: "#A8B9CC",
-  },
-  cpp: {
-    icon: "🚀",
-    color: "#00599C",
-  },
-  csharp: {
-    icon: "🎸",
-    color: "#239120",
-  },
-  php: {
-    icon: "🐘",
-    color: "#777BB4",
-  },
-  ruby: {
-    icon: "💎",
-    color: "#CC342D",
-  },
-  swift: {
-    icon: "🦅",
-    color: "#FA7343",
-  },
-  kotlin: {
-    icon: "🤖",
-    color: "#0095D5",
-  },
-  go: {
-    icon: "🐹",
-    color: "#00ADD8",
-  },
-  rust: {
-    icon: "🦀",
-    color: "#000000",
-  },
-  dart: {
-    icon: "🎯",
-    color: "#0175C2",
-  },
-  html: {
-    icon: "📜",
-    color: "#E34F26",
-  },
-  css: {
-    icon: "🎨",
-    color: "#1572B6",
-  },
-  sql: {
-    icon: "📊",
-    color: "#4479A1",
-  },
-};
+import { languaguesData } from "@/lib/languaguesData";
+import { Repository } from "@/types/repository";
+import { BookOpen, Star, GitFork, AlertCircle, Clock } from "lucide-react";
 
 export const RepositoryCard = ({ repo }: { repo: Repository }) => {
   const formatNumber = (num: number) => {
@@ -166,11 +87,13 @@ export const RepositoryCard = ({ repo }: { repo: Repository }) => {
         <div className="flex items-center justify-between mt-2">
           <div className="flex items-center">
             <span className="text-lg mr-1">
-              {icons?.[repo.language.toLowerCase()]?.icon ?? "💻"}
+              {languaguesData?.[repo.language.toLowerCase()]?.icon ?? "💻"}
             </span>
             <span
               className="text-sm font-medium"
-              style={{ color: icons?.[repo.language.toLowerCase()]?.color }}
+              style={{
+                color: languaguesData?.[repo.language.toLowerCase()]?.color,
+              }}
             >
               {repo.language}
             </span>
