+++ front-end/src/lib/languaguesData.ts
@@ -1,0 +1,70 @@
+export const languaguesData: Record<string, { icon: string; color: string }> = {
+  javascript: {
+    icon: "🌟",
+    color: "#F7DF1E",
+  },
+  typescript: {
+    icon: "🌀",
+    color: "#3178C6",
+  },
+  python: {
+    icon: "🐍",
+    color: "#3776AB",
+  },
+  java: {
+    icon: "☕",
+    color: "#007396",
+  },
+  c: {
+    icon: "🔵",
+    color: "#A8B9CC",
+  },
+  cpp: {
+    icon: "🚀",
+    color: "#00599C",
+  },
+  csharp: {
+    icon: "🎸",
+    color: "#239120",
+  },
+  php: {
+    icon: "🐘",
+    color: "#777BB4",
+  },
+  ruby: {
+    icon: "💎",
+    color: "#CC342D",
+  },
+  swift: {
+    icon: "🦅",
+    color: "#FA7343",
+  },
+  kotlin: {
+    icon: "🤖",
+    color: "#0095D5",
+  },
+  go: {
+    icon: "🐹",
+    color: "#00ADD8",
+  },
+  rust: {
+    icon: "🦀",
+    color: "#000000",
+  },
+  dart: {
+    icon: "🎯",
+    color: "#0175C2",
+  },
+  html: {
+    icon: "📜",
+    color: "#E34F26",
+  },
+  css: {
+    icon: "🎨",
+    color: "#1572B6",
+  },
+  sql: {
+    icon: "📊",
+    color: "#4479A1",
+  },
+};
