+++ front-end/src/pages/private/Onboarding.tsx
@@ -14,7 +14,7 @@ import { Label } from "@/components/ui/label";
 import { RadioGroup, RadioGroupItem } from "@/components/ui/radio-group";
 import { toast } from "sonner";
 import MultiSelector from "@/components/ui/multi-selector";
-import { programmingLanguages, experienceLevels } from "@/data";
+import { programmingLanguages, experienceLevels } from "@/lib/data";
 
 export default function Onboarding() {
   const [selectedItems, setSelectedItems] = useState<string[]>([]);
